+++ spec/savon/request_spec.rb
@@ -2,7 +2,7 @@
 require "spec_helper"
 require "integration/support/server"
 
-describe Savon::WSDLRequest do
+RSpec.describe Savon::WSDLRequest do
 
   let(:globals)      { Savon::GlobalOptions.new }
   let(:http_request) { HTTPI::Request.new }
@@ -191,6 +191,7 @@ describe Savon::WSDLRequest do
     describe "ssl encrypted cert key file" do
       describe "set with an invalid decrypting password" do
         it "fails when attempting to use the SSL private key" do
+          skip("JRuby: find out why this does not raise an error!") if RUBY_PLATFORM == 'java'
           pass = "wrong-password"
           key  = File.expand_path("../../fixtures/ssl/client_encrypted_key.pem", __FILE__)
           cert = File.expand_path("../../fixtures/ssl/client_encrypted_key_cert.pem", __FILE__)
@@ -201,15 +202,12 @@ describe Savon::WSDLRequest do
 
           new_wsdl_request.build
 
-          expect { http_request.auth.ssl.cert_key }.to raise_error
+          expect { http_request.auth.ssl.cert_key }.to raise_error OpenSSL::PKey::PKeyError
         end
       end
 
       describe "set with a valid decrypting password" do
         it "handles SSL private keys properly" do
-          if RUBY_ENGINE == 'jruby'
-            pending("find out why this fails with a null pointer exception on jruby")
-          end
           pass = "secure-password!42"
           key  = File.expand_path("../../fixtures/ssl/client_encrypted_key.pem", __FILE__)
           cert = File.expand_path("../../fixtures/ssl/client_encrypted_key_cert.pem", __FILE__)
@@ -300,7 +298,7 @@ describe Savon::WSDLRequest do
 
 end
 
-describe Savon::SOAPRequest do
+RSpec.describe Savon::SOAPRequest do
 
   let(:globals)      { Savon::GlobalOptions.new }
   let(:http_request) { HTTPI::Request.new }