+++ spec/support/fixture.rb
@@ -9,9 +9,9 @@ class Fixture
       fixtures(type)[fixture] ||= read_file type, fixture
     end
 
-    def response_hash(fixture)
-      @response_hash ||= {}
-      @response_hash[fixture] ||= nori.parse(response(fixture))[:envelope][:body]
+    def full_hash(fixture)
+      @full_hash ||= {}
+      @full_hash[fixture] ||= nori.parse(response(fixture))[:envelope][:body]
     end
 
     TYPES.each do |type, ext|
@@ -21,7 +21,7 @@ class Fixture
   private
 
     def nori
-      Nori.new(:strip_namespaces => true, :convert_tags_to => lambda { |tag| tag.snakecase.to_sym })
+      Nori.new(:strip_namespaces => true, :convert_tags_to => lambda { |tag| Savon::StringUtils.snakecase(tag).to_sym })
     end
 
     def fixtures(type)